+++ src/icon-manifest/supportedExtensions.ts
@@ -1,6 +1,6 @@
 /* tslint:disable max-line-length */
 import { languages } from './languages';
-import { IExtensionCollection, IFileExtension, FileFormat } from '../models/IExtension';
+import { IExtensionCollection, IFileExtension, FileFormat } from '../models';
 
 export const extensions: IExtensionCollection<IFileExtension> = {
   supported: [
