+++ src/models/extensions/FileFormat.ts
@@ -1,0 +1,9 @@
+export enum FileFormat {
+  svg,
+  png,
+  jpg,
+  gif,
+  bmp,
+  tiff,
+  ico,
+}
