+++ src/models/extensions/IExtensionCollection.ts
@@ -1,0 +1,3 @@
+export interface IExtensionCollection<T> {
+  supported: T[];
+}
