+++ src/models/extensions/IFileExtension.ts
@@ -1,0 +1,7 @@
+import { IExtension } from './IExtension';
+import { ILanguage } from '../language';
+
+export interface IFileExtension extends IExtension {
+  filename?: boolean; // set to true if the extension represents the whole file name.
+  languages?: ILanguage[]; // collection of languages associated to the icon.
+}
