+++ src/models/extensions/IFolderExtension.ts
@@ -1,0 +1,3 @@
+import { IExtension } from './IExtension';
+
+export interface IFolderExtension extends IExtension { }
