+++ src/models/iconSchema/IIconPath.ts
@@ -1,0 +1,3 @@
+export interface IIconPath {
+  iconPath: string;
+}
