+++ src/models/iconSchema/ISchemaSettings.ts
@@ -1,0 +1,10 @@
+export interface ISchemaSettings {
+  file: string;
+  folder: string;
+  folderExpanded: string;
+  folderNames: {};
+  folderNamesExpanded: {};
+  fileExtensions: {};
+  fileNames: {};
+  languageIds: {};
+}
