+++ src/models/iconSchema/index.ts
@@ -1,0 +1,3 @@
+export * from './IIconPath';
+export * from './IIconSchema';
+export * from './ISchemaSettings';
