+++ src/models/index.ts
@@ -1,0 +1,7 @@
+export * from './contributions';
+export * from './extensions';
+export * from './iconGenerator';
+export * from './iconSchema';
+export * from './language';
+export * from './settings';
+export * from './vscode';
