+++ src/models/language/ILanguageCollection.ts
@@ -1,7 +1,4 @@
-export interface ILanguage {
-  ids: string | string[];
-  defaultExtension: string; // this is only used for exampleGenerator, so it can know which extension to use.
-}
+import { ILanguage } from './ILanguage';
 
 export interface ILanguageCollection {
   bat: ILanguage;
