+++ src/models/language/index.ts
@@ -1,0 +1,2 @@
+export * from './ILanguage';
+export * from './ILanguageCollection';
