+++ src/models/settings/index.ts
@@ -1,0 +1,2 @@
+export * from './ISettings';
+export * from './IState';
