+++ src/models/vscode/index.ts
@@ -1,0 +1,2 @@
+export *  from './IExtendedWorkspaceConfiguration';
+export *  from './IVSCode';
