+++ src/utils/vscode-extensions.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { IExtendedWorkspaceConfiguration } from '../models/IExtendedWorkspaceConfiguration';
+import { IExtendedWorkspaceConfiguration } from '../models';
 
 export function getConfig() {
   return vscode.workspace.getConfiguration() as IExtendedWorkspaceConfiguration;
