+++ src/welcome/index.ts
@@ -2,8 +2,7 @@ import * as vscode from 'vscode';
 import * as open from 'open';
 import { messages as msg } from '../messages';
 import { getConfig } from '../utils/vscode-extensions';
-import { IState } from '../models/IState';
-import { ISettings } from '../models/ISettings';
+import { IState, ISettings } from '../models';
 import { ISettingsManager, status } from '../settings';
 
 export function manageWelcomeMessage(settingsManager: ISettingsManager) {
